Add singleton/dsiSingleton, check for it in ~distribution
Resolves issue with the test:

 distributions/robust/arithmetic/basics/test_array_accesses1.chpl

+++ modules/internal/ChapelArray.chpl
@@ -818,7 +818,7 @@ module ChapelArray {
     }
 
     proc ~_distribution() {
-      if ! _unowned {
+      if ! _unowned && ! _instance.singleton() {
         on _value {
           // How many domains refer to this distribution?
           var cnt = _instance.destroyDist();
